Add '2jz' to google.txt
2JZ
The 2,997 cc (3.0 L; 182.9 cu in) 2JZ was produced from 1991 (first released in the 1991 Toyota Aristo) to 2007. Cylinder bore and stroke is 86 mm × 86 mm (3.386 in × 3.386 in). VVT-i variable valve timing was first introduced in 1995 starting with the 2JZ-GE, but did not come to the 2JZ-GTE and the US-market 2JZ-GE engines until mid-1997 for the 1998 model year onwards. This engine is not merely a stroked version of the 1JZ (although they share a common bore size, bore pitch, and general architecture), but has a taller block deck and longer connecting rods to accommodate the 14.5 mm (0.57 in) stroke increase.